Removes create-bundle from gulpfile.js, as it is no longer necessary.
Changes index.html to only use phenogrid-byo.js and to use standard HTML and the Phenogrid Javascript API exposed by phenogrid.js.

Fixes _getResourceUrl to detect whether it is dealing with an old /widgets/phenogrid server or a new /node_modules/phenogrid server.

Replace the use of scriptPath with imagePath, which points directly to the /image directory instead of relying upon ../image, which pup_tent can't handle properly.

require jQuery in stickytooltip.js

Eliminate the browserify-shim: usage in package.json, it is no longer necessary. Similarly, the use of the browser: field is not necessary, so has been eliminated.